requirements.txt
circuitsvis==1.41.0
datasets==2.21.0
matplotlib==3.9.2
numpy==1.26.4
python-dotenv==1.0.1
seaborn==0.13.2
tiktoken==0.7.0
torch>=2.4.0
tqdm==4.66.5
transformer-lens==2.4.1
transformers==4.44.2
wandb==0.18.0
